+++ RibbonPreview/Program.cs
@@ -16,6 +16,9 @@ namespace RibbonPreview
         [STAThread]
         static void Main()
         {
+#if Core
+            Application.SetHighDpiMode(HighDpiMode.SystemAware);
+#endif
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
             Application.Run(new MainForm());